moved changes to mobile.css/js
