+++ frontend/index.html
@@ -6,6 +6,7 @@
 <title>Orb</title>
 <link id="theme-link" rel="stylesheet" href="/static/themes/dark.css">
 <link rel="stylesheet" href="/static/style.css">
+<link rel="stylesheet" href="/static/mobile.css">
 <link rel="icon" type="image/png" href="/static/logo.png">
 </head>
 <body>
